framework-comparisons/03-rag-implementation/llamaindex/requirements.txt
llama-index==0.14.13
llama-index-llms-openai==0.6.0
llama-index-embeddings-openai==0.5.1
python-dotenv==1.2.1
langchain-openai==1.1.7
langchain-community==0.4.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
